requirements.txt
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alabaster
altair
anaconda-catalogs
annotated-types
anyio
appdirs==1.4.4
archspec
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
astropy-iers-data
asttokens
async-lru
atomicwrites==1.4.0
attrs
Automat
autopep8
Babel
bcrypt
beautifulsoup4
binaryornot
black
bleach
blinker
bokeh
boltons
botocore
Bottleneck
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
colorcet
comm
conda-content-trust
conda-pack
conda-package-handling
conda-repo-cli
conda_package_streaming
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
dask
dask-expr
datashader
debugpy
decorator
defusedxml
diff-match-patch
dill
distributed
distro
docstring-to-markdown
docutils
et-xmlfile
executing
fastjsonschema
filelock
flake8
Flask
fonttools
frozendict
frozenlist
fsspec
gensim
gitdb
GitPython
greenlet
h11
h5py
HeapDict
holoviews
httpcore
httpx
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jedi
jeepney
jellyfish
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keyring
kiwisolver
lazy-object-proxy
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libmambapy
linkify-it-py
llvmlite
lmdb
locket
lxml
lz4
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.9.2
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
menuinst
mistune
mkl-service==2.4.0
mkl_fft
mkl_random
more-itertools
mpmath
msgpack
multidict
multipledispatch
mypy
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
openpyxl
overrides
packaging
pandas
pandocfilters
panel
param
parsel
parso
partd
pathspec
patsy
pexpect
pickleshare
pillow
pkce
pkginfo
platformdirs
plotly
pluggy
ply
prometheus-client
prompt-toolkit
Protego
protobuf==4.25.3
psutil
ptyprocess
pure-eval
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules==0.2.8
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic
pydantic_core
pydeck
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder==0.4.0
pyodbc
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytest
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
pytoolconfig
pytz
pyviz_comms
PyWavelets
pyxdg
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
referencing
regex
requests
requests-file
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rope
rpds-py
Rtree
ruamel-yaml-conda
ruamel.yaml
ruamel.yaml.clib
s3fs
scikit-image
scikit-learn
scipy
Scrapy
seaborn
SecretStorage
semver
Send2Trash
service-identity
setuptools==75.1.0
sip
six
smart-open
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
stack-data
statsmodels
streamlit
sympy
tables
tabulate
tblib
tenacity
terminado
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
toml
tomli
tomlkit
toolz
tornado
tqdm
traitlets
truststore
Twisted
typing_extensions
tzdata
uc-micro-py
ujson
unicodedata2
Unidecode
urllib3
w3lib
watchdog
wcwidth
webencodings
websocket-client
Werkzeug
whatthepatch
wheel==0.44.0
widgetsnbextension
wrapt
wurlitzer
xarray
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard
